requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
bcrypt
